pyproject.toml
[tool.poetry]
name = "AgentScale"
version = "0.0.1"
description = "AgentScale"
authors = ["Ming XUE <mingxue.mailbox@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/M1n9X/AgentScale"
repository = "https://github.com/M1n9X/AgentScale"
documentation = "https://agentscale.readthedocs.io"
classifiers = ["Development Status :: 3 - Alpha"]

[tool.poetry.urls]
Changelog = "https://github.com/M1n9X/AgentScale/releases"

[tool.poetry.dependencies]
aio-pika = "^9.4.2"
asyncpg = "^0.29.0"
click = ">=8.1.7"
fastapi = "^0.68.0"
hnswlib = "^0.8.0"
ipykernel = "^6.29.4"
isort = "^5.13.2"
nest-asyncio = "^1.6.0"
openai = "^1.35.12"
pgvector = "^0.3.0"
psycopg = { extras = ["binary"], version = ">=3.2.1" }
pypdf = "^4.2.0"
python = ">=3.10,<3.12"
python-consul2 = "^0.1.5"
python-dotenv = "^1.0.1"
pyyaml = ">=6.0.1"
tiktoken = "^0.7.0"
uvicorn = "^0.15.0"

[tool.poetry.dev-dependencies]
Pygments = ">=2.17.2"
coverage = { extras = ["toml"], version = ">=7.4" }
darglint = ">=1.8.1"
flake8 = ">=5.0.4"
flake8-bandit = ">=4.1.1"
flake8-bugbear = ">=23.3.12"
flake8-docstrings = ">=1.7.0"
flake8-rst-docstrings = ">=0.3.0"
furo = ">=2024.1.29"
isort = ">=5.13.2"
mypy = ">=1.9.0"
myst-parser = { version = ">=2.0.0" }
pep8-naming = ">=0.13.3"
pre-commit = ">=3.5.0"
pre-commit-hooks = ">=4.6.0"
pytest = ">=8.1.1"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=5.1.0"
typeguard = ">=4.2.1"
xdoctest = { extras = ["colors"], version = ">=1.1.3" }

[tool.poetry.scripts]
agentscale = "agentscale.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["agentscale", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
